feat: ask user for `enableSecurelyPassphraseCache`
